Add checkUser, checkMember, checkGroup middleware
